requirements.txt
fastapi
grpcio
grpcio-tools
sqlalchemy
uvicorn
